llmv: update golden test case of test/Feature/comdat.ll
Skip new lines in between comdat definitions, as they
are often quite short and fit on one line. This is
analogous to global variable declarations and definitions
which are already printed on a single line with no new
lines in between.

Note, this output differs from opt and Clang, but we
already differ in quite a few other places, so do optimize
for readability rather than compatability in this case.