llvm: update test cases to LLVM 10.0 (part 3) (#4)
* llvm: update test cases for LLVM 10.0

isysroot -> sysroot

Handling for this was implemented by @dannypsnl in llir/llvm#136.

* llvm: sync golden test case of fma.ll to LLVM 10.0

Note, llir/llvm cannot yet handle this round-trip test case as we've
not yet established a canonical representation of NaN values. This is
tracked by llir/llvm#133.

* llvm: update golden test case of DIModuleContext.ll to use sysroot instead of isysroot

* llvm: sync golden test case of compatibility.ll to LLVM 10.0

The main change is the added freeze instruction and the use of local IDs
for unnamed function parameters.